produce a tbl with genomic annotation using a remote parquet file

produce a tbl with genomic annotation using a remote parquet file

Usage

query_osn(con, stub = "gene_info")

Arguments

con

a DBI connection, typically using duckdb, that can be used to query remote parquet files

stub

character(1) tag identifying an annotation resource

Note

DBI::dbListTables is used to query connection for existence of the desired table/view, which is returned if present. Otherwise an httpfs based query is used to produce a view.

Examples

con = DBI::dbConnect(duckdb::duckdb())
query_osn(con, stub="gene_info")
#> # Source:   table<gene_info> [?? x 16]
#> # Database: DuckDB v1.1.3-dev165 [vincentcarey@Darwin 23.6.0:R 4.4.2/:memory:]
#>    tax_id gene_id symbol   locus_tag synonyms    dbXrefs chromosome map_location
#>     <dbl>   <dbl> <chr>    <chr>     <chr>       <chr>   <chr>      <chr>       
#>  1   9606       1 A1BG     -         A1B|ABG|GA… MIM:13… 19         19q13.43    
#>  2   9606       2 A2M      -         A2MD|CPAMD… MIM:10… 12         12p13.31    
#>  3   9606       3 A2MP1    -         A2MP        HGNC:H… 12         12p13.31    
#>  4   9606       9 NAT1     -         AAC1|MNAT|… MIM:10… 8          8p22        
#>  5   9606      10 NAT2     -         AAC2|NAT-2… MIM:61… 8          8p22        
#>  6   9606      11 NATP     -         AACP|NATP1  HGNC:H… 8          8p22        
#>  7   9606      12 SERPINA3 -         AACT|ACT|G… MIM:10… 14         14q32.13    
#>  8   9606      13 AADAC    -         CES5A1|DAC  MIM:60… 3          3q25.1      
#>  9   9606      14 AAMP     -         -           MIM:60… 2          2q35        
#> 10   9606      15 AANAT    -         DSPS|SNAT   MIM:60… 17         17q25.1     
#> # ℹ more rows
#> # ℹ 8 more variables: description <chr>, gene_type <chr>,
#> #   nomenclature_symbol <chr>, nomenclature_name <chr>,
#> #   nomenclature_status <chr>, other_designations <chr>,
#> #   modification_date <chr>, feature_type <chr>
query_osn(con, stub="gene2go")
#> # Source:   table<gene2go> [?? x 8]
#> # Database: DuckDB v1.1.3-dev165 [vincentcarey@Darwin 23.6.0:R 4.4.2/:memory:]
#>    tax_id gene_id go_id  evidence go_qualifier go_description pubmed_id category
#>     <dbl>   <dbl> <chr>  <chr>    <chr>        <chr>          <chr>     <chr>   
#>  1   9606       1 GO:00… IBA      involved_in  immune respon… -         Process 
#>  2   9606       1 GO:00… HDA      located_in   extracellular… 27068509  Compone…
#>  3   9606       1 GO:00… IDA      located_in   extracellular… 3458201   Compone…
#>  4   9606       1 GO:00… TAS      located_in   extracellular… -         Compone…
#>  5   9606       1 GO:00… HDA      located_in   extracellular… 16502470  Compone…
#>  6   9606       1 GO:00… IBA      is_active_in plasma membra… -         Compone…
#>  7   9606       1 GO:00… TAS      located_in   platelet alph… -         Compone…
#>  8   9606       1 GO:00… TAS      located_in   secretory gra… -         Compone…
#>  9   9606       1 GO:00… HDA      located_in   collagen-cont… 27559042  Compone…
#> 10   9606       1 GO:00… HDA      located_in   extracellular… 19056867… Compone…
#> # ℹ more rows
DBI::dbDisconnect(con)